Replace all RBAC adapters by one RbacAdaptor.
