+++ packages/rbac-http/src/index.ts
@@ -1,11 +1,9 @@
-import RbacHttpAdapter from './RbacHttpAdapter';
 import RbacHttpAssignmentAdapter from './adapters/RbacHttpAssignmentAdapter';
 import RbacHttpItemAdapter from './adapters/RbacHttpItemAdapter';
 import RbacHttpItemChildAdapter from './adapters/RbacHttpItemChildAdapter';
 import RbacHttpRuleAdapter from './adapters/RbacHttpRuleAdapter';
 
 export {
-  RbacHttpAdapter,
   RbacHttpAssignmentAdapter,
   RbacHttpItemAdapter,
   RbacHttpItemChildAdapter,
