+++ packages/rbac-in-memory/src/index.ts
@@ -1,11 +1,9 @@
-import RbacInMemoryAdapter from './RbacInMemoryAdapter';
 import RbacInMemoryAssignmentAdapter from './adapters/RbacInMemoryAssignmentAdapter';
 import RbacInMemoryItemAdapter from './adapters/RbacInMemoryItemAdapter';
 import RbacInMemoryItemChildAdapter from './adapters/RbacInMemoryItemChildAdapter';
 import RbacInMemoryRuleAdapter from './adapters/RbacInMemoryRuleAdapter';
 
 export {
-  RbacInMemoryAdapter,
   RbacInMemoryAssignmentAdapter,
   RbacInMemoryItemAdapter,
   RbacInMemoryItemChildAdapter,
