+++ packages/rbac-mongodb/src/index.ts
@@ -1,11 +1,9 @@
-import RbacMongodbAdapter from './RbacMongodbAdapter';
 import RbacMongodbAssignmentAdapter from './adapters/RbacMongodbAssignmentAdapter';
 import RbacMongodbItemAdapter from './adapters/RbacMongodbItemAdapter';
 import RbacMongodbItemChildAdapter from './adapters/RbacMongodbItemChildAdapter';
 import RbacMongodbRuleAdapter from './adapters/RbacMongodbRuleAdapter';
 
 export {
-  RbacMongodbAdapter,
   RbacMongodbAssignmentAdapter,
   RbacMongodbItemAdapter,
   RbacMongodbItemChildAdapter,
