+++ packages/rbac-postgres/src/index.ts
@@ -1,11 +1,9 @@
-import RbacPostgresAdapter from './RbacPostgresAdapter';
 import RbacPostgresAssignmentAdapter from './adapters/RbacPostgresAssignmentAdapter';
 import RbacPostgresItemAdapter from './adapters/RbacPostgresItemAdapter';
 import RbacPostgresItemChildAdapter from './adapters/RbacPostgresItemChildAdapter';
 import RbacPostgresRuleAdapter from './adapters/RbacPostgresRuleAdapter';
 
 export {
-  RbacPostgresAdapter,
   RbacPostgresAssignmentAdapter,
   RbacPostgresItemAdapter,
   RbacPostgresItemChildAdapter,
